+++ src/models/utils/SchemaField.ts
@@ -8,13 +8,6 @@ export default class SchemaField {
     options!:  { [key: string]: any };
 
     constructor(data: { [key: string]: any } = {}) {
-        this.load(data || {});
-    }
-
-    /**
-     * Loads `data` into the field.
-     */
-    load(data: { [key: string]: any }) {
         this.id       = typeof data.id !== 'undefined' ? data.id : '';
         this.name     = typeof data.name !== 'undefined' ? data.name : '';
         this.type     = typeof data.type !== 'undefined' ? data.type : 'text';